Split GradingHistory to separate components for grade, XP, gradingStatus
